requirements.txt
flask==3.1.3
gunicorn==23.0.0
line-bot-sdk==2.4.3
anthropic==0.88.0
supabase==2.9.1
